Close #1002 - xsl transformation of the XML when using SMTPSend step
Javascriptable xslFilepath and contentType properties has been added.

If no source selected, the current outputDocument is used, even if no xsl is defined (act as pretty print).

The xsl is apply to the source results (concatened if more than one) or the outputDocument if no source.

The contentType property can be used to force the content type of the mail.

 * without xsl : text/plain
 * with xsl : text/html
 
 The documentation must be reviewed in #3036

Former-commit-id: 17e3bc7e95d0cf445a8746110270c508993c0e3d
Former-commit-id: 50e822718ba1b9867f98418dfa9fdb8bc31624e8
Former-commit-id: 8164542fce2a00fe9e01f3cdcf5c361a986d5f1f
Former-commit-id: e0a86c1da66e9ea1a5fc2c9886bc8ef01259270f