Close #3188 - Order parallel step response
Element are now ordered.

Execute from attached project the sequence !WhatTimeParallel and the order must be 8/2/4.

Execute the sequence !IteratorParallel with the test case X3 and X10, the order must be the same as the first part of the XML.

Try in an older CEMS version to compare.

A case doesn't work with the Iterator + Parallel case. Only the first output true step is ordered (set ouput false on the Serial of IteratorParallel). We cannot solve this issue with the current sequencer structure.

Former-commit-id: 6d11412de8aadbef1c21165dc50abb9a3c237a4f
Former-commit-id: 093c641f44170c15265c4fd7d4be164390c0665e
Former-commit-id: 9dbc042749ce366332d18a6e6add055f5431a091
Former-commit-id: 8067d37feffc392a3c63e8574d72d7fc772c0b29