Close #3188 - Order parallel step response
Element are now ordered.

Execute from attached project the sequence !WhatTimeParallel and the order must be 8/2/4.

Execute the sequence !IteratorParallel with the test case X3 and X10, the order must be the same as the first part of the XML.

Try in an older CEMS version to compare.

A case doesn't work with the Iterator + Parallel case. Only the first output true step is ordered (set ouput false on the Serial of IteratorParallel). We cannot solve this issue with the current sequencer structure.

Former-commit-id: 5602b6c8f62ad2b0b37853c330e50f31053f544d
Former-commit-id: c0ec6cd052bde6902cbb984e7c9727e9e9d23ead
Former-commit-id: 3ab6a9b6f1bad0e253a8d3e768e33c33b5c6ee07
Former-commit-id: d44f2f980279c17a3cd663bf7ed2f70a1d415a7f