Close #3299 - Add XVNC port Number in Administration summary page .
Done.

The port can be see in the "Java system properties" button on the mail admin page.

This is very specific and doesn't require a dedicated conditional (Linux +DISPLAY env) IHM in the admin page.

A version will be deployed on http://c8o_dev.convertigo.net as soon as the build is available.

Former-commit-id: 06f3b153524eb58bd0d3794414fbcf33094d8c88
Former-commit-id: b391865c88a1f7ae9d04123eb7c954e9db2266f9
Former-commit-id: 54fd59f874e691432c4ded7e606d4b1082e841fd
Former-commit-id: 2e676fabd4fd04f89ed69d5ceb58cfd08a5f665f