Close #3299 - Add XVNC port Number in Administration summary page .
Done.

The port can be see in the "Java system properties" button on the mail admin page.

This is very specific and doesn't require a dedicated conditional (Linux +DISPLAY env) IHM in the admin page.

A version will be deployed on http://c8o_dev.convertigo.net as soon as the build is available.

Former-commit-id: 94c410f4e8ed2f6b133d03444667113b1034ac44
Former-commit-id: 9083b164a8a348275d49fb56e1b6a5a6315d70a1
Former-commit-id: a3f7da1149484fa439d72b72f98b11b812a2e23b
Former-commit-id: de16a51db1c2487dc223a05b78bca34f910a3c77