Close #3082 - Strange behaviors arount SMTP send and Copy steps
Handle the case of empty nodeList to copy with the !XmlCopyStep that cause the step_id attribute.

Handle the case of copy of copy that cause the step_copy attribute.

The Ticket3082 ticket has 2 sequences. One with a bad xpath (cause an empty nodelist) for !XmlCopyStep and another with a valid xpath (generate result).

The darty project should be tested to confirm that.

Done in trunk and 6.2.x branch.

Former-commit-id: afec284acd115e4d94e23810d82cad86a7d5a7a4
Former-commit-id: f8ec34197a2ecc3fdb23f09a7b18c1a4de160739
Former-commit-id: 8c07b745de7844da1c47f65ac791931938da9f68
Former-commit-id: 4b92c9a799a3fe5d67a79d2cd959b0cf36a62d5a