Close #3082 - Strange behaviors arount SMTP send and Copy steps
Handle the case of empty nodeList to copy with the !XmlCopyStep that cause the step_id attribute.

Handle the case of copy of copy that cause the step_copy attribute.

The Ticket3082 ticket has 2 sequences. One with a bad xpath (cause an empty nodelist) for !XmlCopyStep and another with a valid xpath (generate result).

The darty project should be tested to confirm that.

Done in trunk and 6.2.x branch.

Former-commit-id: 14763d378569f17bcd177d843fdb721ae8a13c5b
Former-commit-id: a6b9ef1de322fcd9f1fb1fd4bcb5f1e889d1e036
Former-commit-id: 7e5b70f79920b31d372d0017f114d9bd613edbfc
Former-commit-id: 06fe0c61b83ede763b7ca43dd432450e51be314e