closed #4143 - Fixed complexType name generation (use tagname instead of name)
The trouble here was not really about namespace, but about elements with '''same''' name and from '''different''' namespaces.

I've choosen not to 'merge' types from different namespaces because it doesn't mean anything !
Thus, I've prefered to consider the namespace prefix while generating the element's complexType name.
{{{<xsd:complexType name="root_h-tableType">}}} and {{{<xsd:complexType name="root_f-tableType">}}}

But be aware that the generated schema won't be validated :
{{{
cos-element-consistent : Multiple elements with name 'table', with different types, appear in the model group
}}}
meaning it is forbidden under a given element to have children with same name and different types ! [[BR]]

'''Improvement solution (#4623) :'''
 * the XSDExtractor should be rewritten to use !XmlSchema api
 * it should create a collection of schemas, and
  * or, use referenced schema(s) in XML [[BR]]
  see https://msdn.microsoft.com/en-us/library/ms757863(v=vs.85).aspx
  * or, create a schema for each namespace declared in XML

Former-commit-id: 924060f146f9af845fd6beff9682e7e5540d37cf
Former-commit-id: d8298eb27920aecf3a50d1b73c86630b6e898269
Former-commit-id: fcd93eac821904451c0c50436c97fcb89dcdeabc
Former-commit-id: a8ba529466e4aee25ab7ca0cb1dcfd9cd2a052fa